Refactor FileType Path for Better Modularity
Updated import paths for `FileType` and related interfaces to `src/libs/FileType` to enhance code modularity and organization. This change groups all file type related functionality under a common directory, simplifying imports and improving maintainability.
