+++ src/interfaces/IPrjDocument.ts
@@ -1,4 +1,4 @@
-import { FileType } from 'src/types/FileType/FileType';
+import { FileType } from 'src/libs/FileType/FileType';
 import { FileSubType } from '../types/PrjTypes';
 
 export default interface IPrjDocument {
