+++ src/libs/BlockRenderComponents/TableBlockRenderComponent.ts
@@ -4,8 +4,8 @@ import Lng from 'src/classes/Lng';
 import { Logging } from 'src/classes/Logging';
 import { ILogger } from 'src/interfaces/ILogger';
 import IPrjModel from 'src/interfaces/IPrjModel';
-import { FileType } from 'src/types/FileType/FileType';
-import { FileTypes } from 'src/types/FileType/interfaces/IFileType';
+import { FileType } from 'src/libs/FileType/FileType';
+import { FileTypes } from 'src/libs/FileType/interfaces/IFileType';
 import { PrjSettings } from 'src/types/PrjSettings';
 import RedrawableBlockRenderComponent from './RedrawableBlockRenderComponent';
 import { IProcessorSettings } from '../../interfaces/IProcessorSettings';
