+++ src/libs/FileType/__tests__/FileType.test.ts
