+++ src/libs/FileType/__tests__/IFileType.spec.ts
