+++ src/libs/FileType/interfaces/IFileType.ts
