+++ src/libs/MetadataCache.ts
@@ -4,7 +4,7 @@ import { App, CachedMetadata, TFile } from 'obsidian';
 import { Logging } from 'src/classes/Logging';
 import IMetadataCache from 'src/interfaces/IMetadataCache';
 import { IMetadataCacheEvents } from 'src/interfaces/IMetadataCacheEvents';
-import { FileType } from 'src/types/FileType/FileType';
+import { FileType } from 'src/libs/FileType/FileType';
 import { PrjSettings } from 'src/types/PrjSettings';
 import GenericEvents, { IEvent } from './GenericEvents';
 import Global from '../classes/Global';
