+++ src/models/Data/TaskData.ts
@@ -5,7 +5,7 @@ import IPrjTaskManagement from 'src/interfaces/IPrjTaskManagement';
 import { ITag } from 'src/libs/Tags/interfaces/ITag';
 import { ITags } from 'src/libs/Tags/interfaces/ITags';
 import { Tags } from 'src/libs/Tags/Tags';
-import { FileType } from 'src/types/FileType/FileType';
+import { FileType } from 'src/libs/FileType/FileType';
 import {
     FileSubType,
     Status,
